refactor: 상품 구매 시 재고를 조정하는 로직을 서비스에서 수행하도록 변경
